added test ipynb file to make sure the packeging works, needed to make so changes in packeging
